requirements.txt
alabaster==0.7.12
asgiref==3.2.3     
astroid==2.3.3     
atomicwrites==1.3.0
attrs==19.3.0      
Babel==2.8.0       
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
django-heroku==0.3.1
django-background-tasks==1.2.5
django-compat==1.0.15
django-extensions==2.2.8
django-sslserver==0.22
docopt==0.6.2
docutils==0.16
entrypoints==0.3
futures==3.1.1
google==2.0.3
google-api-core==1.16.0
google-auth==1.11.0
google-cloud-bigquery==1.24.0
google-cloud-core==1.3.0
google-cloud-storage==1.25.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
googletrans==2.4.0
goslate==1.5.1
gunicorn==20.0.4
idna==2.8
imagesize==1.2.0
ipykernel==5.1.4
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
Js2Py==0.67
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lxml==4.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.2.0
nbconvert==5.6.1
nbformat==5.0.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
numpy==1.18.1
packaging==20.1
pandocfilters==1.4.2
parso==0.6.1
pickleshare==0.7.5
Pillow==7.0.0
pipwin==0.4.6
pluggy==0.13.1
prometheus-client==0.7.1
protobuf==3.11.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
pyjsparser==2.7.1
pylint==2.4.4
PyMySQL==0.9.3
pyparsing==2.4.6
PyPrind==2.11.2
pyrsistent==0.15.7
pytest==5.3.5
python-dateutil==2.8.1
PythonTurtle==0.3.1
pytrends==4.7.2
pytz==2019.3
pyzmq==18.1.1
qtconsole==4.6.0
regex==2020.1.8
requests==2.22.0
robobrowser==0.5.3
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
snowballstemmer==2.0.0
soupsieve==1.9.5
SpeechRecognition==3.8.1
Sphinx==2.4.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
stopwords==0.1.3
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
tabulate==0.8.6
terminado==0.8.3
testpath==0.4.4
textblob==0.15.3
tornado==6.0.3
traitlets==4.3.3
tzlocal==2.0.0
urllib3==1.25.8
virtualenv==16.7.9
virtualenvwrapper-win==1.2.5
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
whitenoise==5.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
yagmail==0.11.224
